src/api/requirements.txt
flask
requests
python-dotenv
beautifulsoup4
youtube-dl
flask_restplus